Fix: fetch 코드로 변경
